bug fix in PCL
