+++ CoreTweet.Shared/CoreTweet.Shared.projitems
@@ -23,6 +23,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Internal\Request.Async.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Internal\Request.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Internal\SecurityUtils.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Internal\SimpleProgress.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Internal\TokensBase.Async.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Internal\TokensBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Internal\InternalUtils.cs" />
